optimized create_bundle_attributes
